bq_agent_system/requirements.txt
google-cloud-geminidataanalytics>=0.1.0
google-cloud-storage>=2.10.0
google-cloud-dataplex>=2.0.0
google-cloud-bigquery>=3.11.0
streamlit>=1.28.0
pydantic>=2.0.0
